# src/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "cybersource",
    "contact": {},
    "version": "1.0"
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "variables": {}
    }
  ],
  "paths": {
    "/api/v1/aggregators/register": {
      "post": {
        "tags": [
          "aggregators"
        ],
        "summary": "register aggregator",
        "operationId": "registeraggregator",
        "parameters": [],
        "requestBody": {
          "description": "",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/registeraggregatorrequest"
              },
              "example": {
                "companyName": "The Company1.0",
                "phoneNumber": "09065734215",
                "businessYears": 9,
                "email": "theemail14@email.com",
                "websiteLink": "www.the-website.com",
                "accountNumber": "0763239709",
                "address": "No 2 Luke Skywall Street",
                "primaryContactName": "The Primary Name",
                "primaryContactTelephone": "08984347322",
                "primaryContactPhoneNumber": "08033473333",
                "primaryContactEmail": "theprimary@email.com",
                "secondaryContactName": "The Secondary Name",
                "secondaryContactTelephone": "08044334444",
                "secondaryContactPhoneNumber": "08055544555",
                "secondaryContactEmail": "thesecondary@email.com"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "",
            "headers": {}
          }
        },
        "deprecated": false
      }
    },
...
    "/api/v1/aggregators/1/approve": {
      "patch": {
        "tags": [
          "aggregators"
        ],
        "summary": "approve aggregator",
        "operationId": "approveaggregator",
        "parameters": [
          {
            "name": "Authorization",
            "in": "header",
            "description": "",
            "required": true,
            "style": "simple",
            "schema": {
              "type": "string",
              "example": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJ1c2VybmFtZSI6ImVuZWphayIsInJvbGUiOiJhZG1pbiIsImlzVmVyaWZpZWQiOnRydWUsImlzQWN0aXZlIjp0cnVlLCJwZW5kaW5nIjpmYWxzZSwiaWF0IjoxNTkyMjg4NzAxLCJleHAiOjE1OTIzNzUxMDF9.evPBJKXP8cI9TY6dHshEWEZnNUhwCJfQ9X3XkpZsEco"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "headers": {}
          }
        },
        "deprecated": false,
        "security": []
      }
    },
...
    "/api/v1/aggregators/reset_password": {
      "post": {
        "tags": [
          "aggregators"
        ],
        "summary": "reset aggregator password",
        "operationId": "resetaggregatorpassword",
        "parameters": [],
        "requestBody": {
          "description": "",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/resetaggregatorpasswordrequest"
              },
              "example": {
                "email": "valid_aggregator_email"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "",
            "headers": {}
          }
        },
        "deprecated": false
      }
    },
...
  },
  "components": {
    "schemas": {
      "registeraggregatorrequest": {
        "title": "registeraggregatorrequest",
        "required": [
          "companyName",
          "phoneNumber",
          "businessYears",
          "email",
          "websiteLink",
          "accountNumber",
          "address",
          "primaryContactName",
          "primaryContactTelephone",
          "primaryContactPhoneNumber",
          "primaryContactEmail",
          "secondaryContactName",
          "secondaryContactTelephone",
          "secondaryContactPhoneNumber",
          "secondaryContactEmail"
        ],
        "type": "object",
        "properties": {
          "companyName": {
            "type": "string"
          },
          "phoneNumber": {
            "type": "string"
          },
          "businessYears": {
            "type": "integer",
            "format": "int32"
          },
          "email": {
            "type": "string"
          },
          "websiteLink": {
            "type": "string"
          },
          "accountNumber": {
            "type": "string"
          },
          "address": {
            "type": "string"
          },
          "primaryContactName": {
            "type": "string"
          },
          "primaryContactTelephone": {
            "type": "string"
          },
          "primaryContactPhoneNumber": {
            "type": "string"
          },
          "primaryContactEmail": {
            "type": "string"
          },
          "secondaryContactName": {
            "type": "string"
          },
          "secondaryContactTelephone": {
            "type": "string"
          },
          "secondaryContactPhoneNumber": {
            "type": "string"
          },
          "secondaryContactEmail": {
            "type": "string"
          }
        },
        "example": {
          "companyName": "The Company1.0",
          "phoneNumber": "09065734215",
          "businessYears": 9,
          "email": "theemail14@email.com",
          "websiteLink": "www.the-website.com",
          "accountNumber": "0763239709",
          "address": "No 2 Luke Skywall Street",
          "primaryContactName": "The Primary Name",
          "primaryContactTelephone": "08984347322",
          "primaryContactPhoneNumber": "08033473333",
          "primaryContactEmail": "theprimary@email.com",
          "secondaryContactName": "The Secondary Name",
          "secondaryContactTelephone": "08044334444",
          "secondaryContactPhoneNumber": "08055544555",
          "secondaryContactEmail": "thesecondary@email.com"
        }
      },
...
      "resetaggregatorpasswordrequest": {
        "title": "resetaggregatorpasswordrequest",
        "required": [
          "email"
        ],
        "type": "object",
        "properties": {
          "email": {
            "type": "string"
          }
        },
        "example": {
          "email": "valid_aggregator_email"
        }
      },
...
    }
  },
...
}